+++ frontend/lib/database.ts
@@ -1,2 +1,2 @@
 import 'server-only';
-export * from '../../db/client';
+export * from '@/db/client';
