+++ frontend/package.json
@@ -51,6 +51,7 @@
     "postcss": "^8",
     "prettier": "^3.5.3",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5"
+    "typescript": "^5",
+    "drizzle-kit": "^0.27.0"
   }
 }
