+++ package.json
@@ -15,8 +15,6 @@
     "deploy:upgradeable": "hardhat run scripts/deploy-upgradeable.ts",
     "upgrade:contract": "hardhat run scripts/upgrade-contract.ts",
     "test:upgradeable": "hardhat test test/unit/UpgradeableGame.test.ts",
-    "db:migrate": "drizzle-kit push",
-    "db:generate": "drizzle-kit generate:pg",
     "db:clear": "./migrations/clean-db.sh",
     "fetch:game": "ts-node scripts/fetch-game.ts",
     "fetch:team": "ts-node scripts/fetch-team.ts",
@@ -33,7 +31,6 @@
     "@uniswap/lib": "^4.0.1-alpha",
     "create-onchain": "^0.0.24",
     "dotenv": "^17.2.1",
-    "drizzle-orm": "^0.35.3",
     "eciesjs": "^0.4.15",
     "elliptic": "^6.6.1",
     "elo-lib": "^0.0.3",
@@ -62,7 +59,6 @@
     "@types/mocha": "^10.0.8",
     "@types/node": "^22.8.5",
     "chai": "^4.5.0",
-    "drizzle-kit": "^0.27.0",
     "eslint": "^9",
     "ethereum-waffle": "^3.4.4",
     "forge-std": "foundry-rs/forge-std#v1.9.4",
