+++ test/integration/VerifyPlayedGame.test.ts
@@ -3,7 +3,7 @@
 // ============================================================================
 
 import { expect } from 'chai';
-import { pool } from '../../db/client';
+import { pool } from '../../frontend/db/pool';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
